+++ server/actions/webhook.js
@@ -17,6 +17,34 @@ export default function webhook(req, res, next) {
     .then(next, next);
   }
 
+  if (_.get(req, "body.topic") === "contact.created") {
+    const lead = _.get(req, "body.data.item");
+    return Batcher.getHandler("webhook_leads", {
+      ctx: req.hull,
+      options: {
+        maxSize: process.env.NOTIFY_BATCH_HANDLER_SIZE || 100,
+        throttle: process.env.NOTIFY_BATCH_HANDLER_THROTTLE || 30000
+      }
+    })
+    .setCallback(leads => req.hull.enqueue("saveLeads", { leads }))
+    .addMessage(lead)
+    .then(next, next);
+  }
+
+  if (_.get(req, "body.topic") === "contact.signed_up") {
+    const lead = _.get(req, "body.data.item");
+    return Batcher.getHandler("webhook_contact.signed_up", {
+      ctx: req.hull,
+      options: {
+        maxSize: process.env.NOTIFY_BATCH_HANDLER_SIZE || 100,
+        throttle: process.env.NOTIFY_BATCH_HANDLER_THROTTLE || 30000
+      }
+    })
+    .setCallback(users => req.hull.enqueue("saveUsers", { users }))
+    .addMessage(lead)
+    .then(next, next);
+  }
+
   return Batcher.getHandler("webhook_events", {
     ctx: req.hull,
     options: {