+++ server/jobs/fetch-users.js
@@ -7,7 +7,7 @@ import fetchAllUsers from "./fetch-all-users";
 import handleRateLimitError from "../lib/handle-rate-limit-error";
 
 export default function fetchUsers(ctx, payload = {}) {
-  const { intercomAgent } = ctx.shipApp;
+  const { intercomAgent } = ctx.service;
   const { count = 50, page = 1 } = payload;
   let { last_updated_at } = payload;