+++ server/jobs/handle-bulk.js
@@ -5,7 +5,7 @@ import handleRateLimitError from "../lib/handle-rate-limit-error";
 
 export default function handleBulk(ctx, payload) {
   const { id, users, attempt = 1 } = payload;
-  const { syncAgent, intercomAgent } = ctx.shipApp;
+  const { syncAgent, intercomAgent } = ctx.service;
   return intercomAgent.getJob(id)
     .then(({ isCompleted, hasErrors }) => {
       if (isCompleted) {