+++ server/jobs/send-users.js
@@ -9,7 +9,7 @@ import handleRateLimitError from "../lib/handle-rate-limit-error";
  */
 export default function sendUsers(ctx, payload) {
   const { users, mode = "bulk" } = payload;
-  const { syncAgent, intercomAgent } = ctx.shipApp;
+  const { syncAgent, intercomAgent } = ctx.service;
 
   ctx.client.logger.debug("sendUsers.preFilter", users.length);
   const usersToSave = syncAgent.getUsersToSave(users);