+++ server/router/app.js
@@ -4,7 +4,7 @@ import cors from "cors";
 import { notifHandler, responseMiddleware } from "hull/lib/utils";
 
 import appMiddleware from "../lib/middleware/app-middleware";
-import requireConfiguration from "../lib/require-configuration";
+import requireConfiguration from "../lib/middleware/require-configuration";
 import * as notifHandlers from "./../notif-handlers";
 import * as actions from "./../actions";
 
@@ -39,6 +39,8 @@ export default function AppRouter(): Router {
   router.use("/intercom", requireConfiguration, actions.webhook, responseMiddleware());
 
   router.post("/sync", requireConfiguration, actions.sync, responseMiddleware());
+
+  router.post("/fetch-leads", requireConfiguration, actions.fetchLeads, responseMiddleware());
 
   router.get("/schema/user_fields", cors(), requireConfiguration, actions.fields);